move plugin types out of plugins folder
