+++ source/lib/pluginTypes.ts
